+++ packages/leva/plugins/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "dist/leva-leva-plugins.cjs.js"
+  "main": "dist/leva-plugins.cjs.js"
 }
