+++ packages/leva/utilities/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "dist/leva-leva-utilities.cjs.js"
+  "main": "dist/leva-utilities.cjs.js"
 }
